0523-16h

+++ accounts/views.py
@@ -4,8 +4,9 @@ from django.conf import settings
 from rest_framework.decorators import api_view, authentication_classes
 from rest_framework.authentication import TokenAuthentication
 from faker import Faker
+from fp.bank_img import BANK_IMAGE_URL_DICT
 import random
-from savedata.models import DepositProducts, AnnuityProducts, creditLoanProducts
+from savedata.models import DepositProducts, DepositOptions, AnnuityProducts, creditLoanProducts
 from .models import User
 from savedata.serializers import DepositProductsSerializer, DepositOptionsSerializer
 from .serializers import UserSerializer
@@ -27,8 +28,18 @@ def profile(request):
     if request.method == 'GET':
         user_serializer = UserSerializer(user)
         user_info = user_serializer.data
+        like_product = []
         products = user.product
         pro_serializer = DepositProductsSerializer(products, many=True)
+        for product in pro_serializer.data:
+            kor_co_nm = product.get('kor_co_nm')
+            fin_prdt_cd = product.get('fin_prdt_cd')
+            fin_co_no = product.get('fin_co_no')
+            product['imgUrl'] = BANK_IMAGE_URL_DICT.get(kor_co_nm, '/assets/icons/banks/default-logo.svg')
+            options = DepositOptions.objects.filter(fin_co_no=fin_co_no, fin_prdt_cd=fin_prdt_cd).order_by('-intr_rate_type_nm','-intr_rate2')[0]
+            opt_serializer = DepositOptionsSerializer(options)
+            product['intr_rate2'] = opt_serializer.data.get("intr_rate2")
+            like_product.append(product)
         data = {
             'user': {
                 'id':user_info.get('id'),
@@ -36,7 +47,7 @@ def profile(request):
                 'birth':user_info.get('birth'),
                 'email':user_info.get('email')
             },
-            'product':pro_serializer.data,
+            'product':like_product,
         }
         return Response(data, status=status.HTTP_200_OK)
     # nickname 수정
@@ -51,7 +62,7 @@ def profile(request):
 def reset_pw(request):
     # 이메일로 유저찾기
     if request.method == 'POST':
-        user = User.objects.get(email=request.POST['email'])
+        user = User.objects.get(email=request.data['email'])
         user_serializer = UserSerializer(user)
         return Response(user_serializer.data, status=status.HTTP_200_OK)
     # 비밀번호 변경
@@ -69,10 +80,14 @@ def reset_pw(request):
 def delete_product(request, product_pk):
     # 예/적금 찜한 상품 삭제
     user = User.objects.get(username=request.user.username)
-    product = user.product.get(pk=product_pk)
-    product.delete()
+    product = user.product.filter(pk=product_pk)
+    if product.exists():
+        product.delete()
+        msg = {'message':f"{product_pk} 상품 삭제 완료"}
+    else:
+        msg = {'message':f"{product_pk} 상품 없음"}
     
-    Response({'message':f"{product_pk} 상품 삭제 완료"}, status=status.HTTP_204_NO_CONTENT)
+    return Response(msg, status=status.HTTP_200_OK)
 
 
 
@@ -89,9 +104,11 @@ def fake_user(request):
         birth = fake.date_between(start_date='-90y', end_date='-19y')
         gender = fake.pyint(min_value=0, max_value=1)
         crdt_grad = fake.pyint(min_value=10, max_value=999)
-        salary = fake.pyint(min_value=2500, max_value=20000)
-        m_consumption = fake.pyint(min_value=0, max_value=300)
-        creditcard = fake.pyint(min_value=0, max_value=5)
+        salary = fake.pyint(min_value=1800, max_value=24000)
+        m_consumption = fake.pyint(min_value=50, max_value=1000)
+        asset = fake.pyint(min_value=-5000, max_value=100000)
+        real_estate = fake.pyint(min_value=0, max_value=1)
+        invest_tendency = fake.pyint(min_value=0, max_value=5)
         save_user = {
             'username':email.split('@')[0],
             'password':make_password("test_password"),
@@ -102,7 +119,9 @@ def fake_user(request):
             'crdt_grad':crdt_grad,
             'salary':salary,
             'm_consumption':m_consumption,
-            'creditcard':creditcard,
+            'asset':asset,
+            'real_estate':real_estate,
+            'invest_tendency':invest_tendency,
         }
         serializer = UserSerializer(data = save_user)
         if serializer.is_valid(raise_exception=True):
